chore(derive): hoist attributes queue test utils
