+++ crates/derive/src/stages/test_utils/mod.rs
@@ -7,11 +7,6 @@ pub use batch_queue::MockBatchQueueProvider;
 mod batch_stream;
 pub use batch_stream::MockBatchStreamProvider;
 
-mod attributes_queue;
-pub use attributes_queue::{
-    new_attributes_provider, MockAttributesBuilder, MockAttributesProvider,
-};
-
 mod frame_queue;
 pub use frame_queue::MockFrameQueueProvider;
