+++ crates/derive/src/test_utils/stages/mod.rs
@@ -1,0 +1,6 @@
+//! Test Utilities for `kona-derive`'s stages.
+
+pub mod attributes_queue;
+pub use attributes_queue::{
+    new_test_attributes_provider, TestAttributesBuilder, TestAttributesProvider,
+};
